LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/feature-policy/feature-policy-frame-policy-timing.https.sub-expected.txt

Blocked access to external URL https://www1.localhost:9443/feature-policy/feature-policy-frame-policy-timing-iframe-camera.https.sub.html


FAIL allow attr timing test same origin assert_equals: expected "NotAllowedError" but got "SecurityError"
FAIL allow attr timing test diff origin assert_equals: expected "NotAllowedError" but got "timeout"
